% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code_run.R
\name{code_run}
\alias{code_run}
\title{Run the Bayesian SDDM model with HMC}
\usage{
code_run(
  save_location,
  dat,
  cluster_label,
  Anxiety_status,
  sub_type,
  sample_size,
  file_name,
  prep_data_file_name,
  range_main_eff,
  range_p,
  range_d,
  range_stop_prob,
  range_rand_eff,
  range_rand_g_l,
  range_rand_g_r,
  range_rand_b_l,
  range_rand_b_r,
  penal_param.int,
  delta_prime.int,
  stop_param.int,
  prob_param.int,
  rand_param_g_l.int,
  rand_param_g_r.int,
  rand_param_b_l.int,
  rand_param_b_r.int,
  nu_d,
  kappa,
  a,
  b,
  CA_threshold,
  nknots,
  m,
  scale,
  SSD_min,
  upper_bound,
  L,
  leapmax,
  thin,
  nparall,
  nhmc,
  intercept
)
}
\arguments{
\item{save_location}{Directory to save all model outputs}

\item{dat}{Input data frame with stop-signal and go-trial information}

\item{cluster_label}{Cluster label used to subset the data}

\item{Anxiety_status}{Group label used to subset the data}

\item{sub_type}{Subtype category used to subset the data}

\item{sample_size}{Number of subjects to sample from the subset}

\item{file_name}{File path to save the model output (RDS)}

\item{prep_data_file_name}{File path to save the preprocessed data (RDS)}

\item{range_main_eff}{Prior range for main effects}

\item{range_p, range_d, range_stop_prob}{Prior ranges for penalty, delta, and stop probability}

\item{range_rand_eff, range_rand_g_l, range_rand_g_r, range_rand_b_l, range_rand_b_r}{Ranges for random effects}

\item{penal_param.int}{Initial penalty parameters}

\item{delta_prime.int}{Initial delta prime parameters}

\item{stop_param.int}{Initial stop parameters}

\item{prob_param.int}{Initial probability parameters}

\item{rand_param_g_l.int, rand_param_g_r.int, rand_param_b_l.int, rand_param_b_r.int}{Initial random parameters}

\item{nu_d}{Degrees of freedom for random effect prior}

\item{kappa}{Shape parameter for inverse gamma prior}

\item{a, b}{Hyperparameters for the prior}

\item{CA_threshold}{Threshold for subject-level choice accuracy}

\item{nknots}{Number of spline knots}

\item{m}{Number of basis functions}

\item{scale}{Rescaling factor}

\item{SSD_min}{Minimum Stop Signal Delay}

\item{upper_bound}{Upper bound for integration}

\item{L}{Number of leapfrog steps}

\item{leapmax}{Maximum leap size}

\item{thin}{Thinning interval}

\item{nparall}{Number of parallel threads}

\item{nhmc}{Number of HMC iterations}

\item{intercept}{Logical. Whether to include intercept terms in spline}
}
\value{
Saves two RDS files: one with preprocessed data and another with the model output
}
\description{
This function prepares a dataset, initializes priors, loads necessary C++ functions,
and runs a Bayesian Semi-parametric Drift Diffusion Model (SSDDM) using Hamiltonian Monte Carlo (HMC).
}
